chore: organize comment-related files
